+++ examples/for-tests-react-16/src/AppWithReactDomRouterV5.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-domv5";
-import { getSuperTokensRoutesForReactRouterDom } from "supertokens-auth-react/prebuiltui";
+import { getSuperTokensRoutesForReactRouterDom } from "supertokens-auth-react/ui";
 import { SessionAuth } from "supertokens-auth-react/recipe/session";
 import { SignInAndUp } from "supertokens-auth-react/recipe/emailpassword/prebuiltui";
 import { BaseComponent, Home, Contact, Dashboard, DashboardNoAuthRequired } from "./App";
@@ -11,34 +11,34 @@ import { EmailPasswordPreBuiltUI } from "supertokens-auth-react/recipe/emailpass
 import { PasswordlessPreBuiltUI } from "supertokens-auth-react/recipe/passwordless/prebuiltui";
 import { ThirdPartyPreBuiltUI } from "supertokens-auth-react/recipe/thirdparty/prebuiltui";
 
-const authRecipe = window.localStorage.getItem("authRecipe") || "emailpassword";
-const emailVerificationMode = window.localStorage.getItem("mode") || "OFF";
-
-let recipePreBuiltUIList = [EmailPasswordPreBuiltUI];
-if (authRecipe === "thirdparty") {
-    recipePreBuiltUIList = [ThirdPartyPreBuiltUI];
-} else if (authRecipe === "emailpassword") {
-    recipePreBuiltUIList = [EmailPasswordPreBuiltUI];
-} else if (authRecipe === "both") {
-    recipePreBuiltUIList = [EmailPasswordPreBuiltUI, ThirdPartyPreBuiltUI];
-} else if (authRecipe === "thirdpartyemailpassword") {
-    recipePreBuiltUIList = [ThirdPartyEmailPasswordPreBuiltUI];
-} else if (authRecipe === "passwordless") {
-    recipePreBuiltUIList = [PasswordlessPreBuiltUI];
-} else if (authRecipe === "thirdpartypasswordless") {
-    recipePreBuiltUIList = [ThirdPartyPasswordlessPreBuiltUI];
-}
-if (emailVerificationMode !== "OFF") {
-    recipePreBuiltUIList.push(EmailVerificationPreBuiltUI);
-}
-
 function AppWithReactDomRouter(props) {
+    const authRecipe = window.localStorage.getItem("authRecipe") || "emailpassword";
+    const emailVerificationMode = window.localStorage.getItem("mode") || "OFF";
+
+    let recipePreBuiltUIList = [EmailPasswordPreBuiltUI];
+    if (authRecipe === "thirdparty") {
+        recipePreBuiltUIList = [ThirdPartyPreBuiltUI];
+    } else if (authRecipe === "emailpassword") {
+        recipePreBuiltUIList = [EmailPasswordPreBuiltUI];
+    } else if (authRecipe === "both") {
+        recipePreBuiltUIList = [EmailPasswordPreBuiltUI, ThirdPartyPreBuiltUI];
+    } else if (authRecipe === "thirdpartyemailpassword") {
+        recipePreBuiltUIList = [ThirdPartyEmailPasswordPreBuiltUI];
+    } else if (authRecipe === "passwordless") {
+        recipePreBuiltUIList = [PasswordlessPreBuiltUI];
+    } else if (authRecipe === "thirdpartypasswordless") {
+        recipePreBuiltUIList = [ThirdPartyPasswordlessPreBuiltUI];
+    }
+    if (emailVerificationMode !== "OFF") {
+        recipePreBuiltUIList.push(EmailVerificationPreBuiltUI);
+    }
+
     return (
         <div className="App">
             <Router>
                 <BaseComponent>
                     <Switch>
-                        {getSuperTokensRoutesForReactRouterDom(require("react-router-domv5", recipePreBuiltUIList))}
+                        {getSuperTokensRoutesForReactRouterDom(require("react-router-domv5"), recipePreBuiltUIList)}
                         <Route exact path="/">
                             <Home />
                         </Route>