+++ lib/ts/claims/primitiveArrayClaim.ts
@@ -1,0 +1,32 @@
+import { PrimitiveArrayClaim as PrimitiveArrayClaimWebJS } from "supertokens-web-js/recipe/session";
+
+import type { ValidationFailureCallback } from "../types";
+import type { PrimitiveArrayClaimConfig } from "supertokens-web-js/recipe/session";
+
+export class PrimitiveArrayClaim<T> extends PrimitiveArrayClaimWebJS<T> {
+    constructor(
+        config: PrimitiveArrayClaimConfig & {
+            onFailureRedirection?: ValidationFailureCallback;
+            showAccessDeniedOnFailure?: boolean;
+        }
+    ) {
+        super(config);
+
+        const validatorsWithCallbacks: { [key: string]: any } = { ...this.validators };
+
+        for (const key in validatorsWithCallbacks) {
+            const validator = validatorsWithCallbacks[key];
+            validatorsWithCallbacks[key as keyof PrimitiveArrayClaimWebJS<T>["validators"]] = (
+                ...args: Parameters<typeof validator>
+            ) => {
+                return {
+                    ...validator(...args),
+                    onFailureRedirection: config.onFailureRedirection,
+                    showAccessDeniedOnFailure: config.showAccessDeniedOnFailure,
+                };
+            };
+        }
+
+        this.validators = validatorsWithCallbacks as PrimitiveArrayClaimWebJS<T>["validators"];
+    }
+}