+++ lib/ts/recipe/emailverification/recipe.tsx
@@ -24,7 +24,6 @@ import { PostSuperTokensInitCallbacks } from "supertokens-web-js/utils/postSuper
 import { SessionClaimValidatorStore } from "supertokens-web-js/utils/sessionClaimValidatorStore";
 
 import { SSR_ERROR } from "../../constants";
-import { saveInvalidClaimRedirectPathInContext } from "../../utils";
 import RecipeModule from "../recipeModule";
 
 import { DEFAULT_VERIFY_EMAIL_PATH } from "./constants";
@@ -49,17 +48,9 @@ export default class EmailVerification extends RecipeModule<
 > {
     static instance?: EmailVerification;
     static RECIPE_ID = "emailverification";
+
     static EmailVerificationClaim = new EmailVerificationClaimClass(
-        () => EmailVerification.getInstanceOrThrow().webJSRecipe,
-        async (userContext: any) => {
-            const recipe = EmailVerification.getInstanceOrThrow();
-            if (recipe.config.mode === "REQUIRED") {
-                saveInvalidClaimRedirectPathInContext(
-                    userContext,
-                    await recipe.getRedirectUrl({ action: "VERIFY_EMAIL" })
-                );
-            }
-        }
+        () => EmailVerification.getInstanceOrThrow().webJSRecipe
     );
 
     constructor(
@@ -69,9 +60,8 @@ export default class EmailVerification extends RecipeModule<
         super(config);
 
         PostSuperTokensInitCallbacks.addPostInitCallback(() => {
-            SessionClaimValidatorStore.addClaimValidatorFromOtherRecipe(
-                EmailVerification.EmailVerificationClaim.validators.isVerified(10)
-            );
+            const isVerifiedValidator = EmailVerification.EmailVerificationClaim.validators.isVerified(10);
+            SessionClaimValidatorStore.addClaimValidatorFromOtherRecipe(isVerifiedValidator);
         });
     }