+++ lib/ts/types.ts
@@ -18,6 +18,7 @@ import type { NormalisedConfig as NormalisedRecipeModuleConfig } from "./recipe/
 import type { TranslationFunc, TranslationStore } from "./translation/translationHelpers";
 import type { ComponentClass, PropsWithChildren } from "react";
 import type { CreateRecipeFunction as CreateRecipeFunctionWebJS } from "supertokens-web-js/lib/build/types";
+import type { SessionClaimValidator as SessionClaimValidatorWebJS } from "supertokens-web-js/recipe/session";
 import type { CookieHandlerInput } from "supertokens-web-js/utils/cookieHandler/types";
 import type NormalisedURLDomain from "supertokens-web-js/utils/normalisedURLDomain";
 import type NormalisedURLPath from "supertokens-web-js/utils/normalisedURLPath";
@@ -25,7 +26,16 @@ import type { WindowHandlerInput } from "supertokens-web-js/utils/windowHandler/
 
 export type GetRedirectionURLContext = {
     action: "TO_AUTH";
-    showSignIn: boolean | undefined;
+    showSignIn?: boolean;
+};
+
+export type ValidationFailureCallback =
+    | (({ userContext, reason }: { userContext: any; reason: any }) => Promise<string | undefined> | string | undefined)
+    | undefined;
+
+export type SessionClaimValidator = SessionClaimValidatorWebJS & {
+    showAccessDeniedOnFailure?: boolean;
+    onFailureRedirection?: ValidationFailureCallback;
 };
 
 /*