+++ lib/ts/utils.ts
@@ -452,23 +452,3 @@ export const useOnMountAPICall = <T>(
         throw error;
     }
 };
-
-export function saveInvalidClaimRedirectPathInContext(userContext: any, invalidClaimRedirectPath: string) {
-    if (userContext["_default"] === undefined) {
-        userContext["_default"] = {};
-    }
-    if (userContext["_default"].redirectPath === undefined) {
-        userContext["_default"] = {
-            ...userContext["_default"],
-            invalidClaimRedirectPath,
-        };
-    }
-}
-
-export function popInvalidClaimRedirectPathFromContext(userContext: any) {
-    const res = userContext["_default"]?.invalidClaimRedirectPath;
-    if (res !== undefined) {
-        delete userContext["_default"].invalidClaimRedirectPath;
-    }
-    return res;
-}