+++ test/unit/recipe/session/sessionAuth.test.tsx
@@ -7,7 +7,7 @@ import SessionAuth from "../../../../lib/ts/recipe/session/sessionAuth";
 import SessionContext from "../../../../lib/ts/recipe/session/sessionContext";
 import { SessionContextType } from "../../../../lib/ts/recipe/session";
 import { PrimitiveClaim, SessionClaim, useClaimValue } from "../../../../lib/ts/recipe/session";
-import { saveInvalidClaimRedirectPathInContext } from "../../../../lib/ts/utils";
+import * as utils from "supertokens-web-js/utils";
 
 const TestClaim: SessionClaim<string> = new PrimitiveClaim({
     id: "st-test-claim",
@@ -74,6 +74,10 @@ jest.spyOn(SuperTokens, "getInstanceOrThrow").mockImplementation(
         } as any)
 );
 jest.spyOn(Session, "getInstanceOrThrow").mockImplementation(() => MockSession as any);
+
+jest.spyOn(utils, "getGlobalClaimValidators").mockReturnValue([
+    { id: "st-test-claim", onFailureRedirection: () => "/test-redirect" },
+] as any[]);
 
 describe("SessionAuth", () => {
     beforeEach(() => {
@@ -498,10 +502,12 @@ describe("SessionAuth", () => {
                 },
             };
 
-            // when
-            MockSession.validateClaims.mockImplementationOnce(({ userContext }) => {
-                saveInvalidClaimRedirectPathInContext(userContext, "/test-redirect");
-                return [{ id: "st-test-claim" }];
+            setMockResolves({
+                doesSessionExist: true,
+                accessTokenPayload: {},
+                userId: "mock-id",
+                invalidClaims: [{ validatorId: "st-test-claim", reason: "test-reason" }],
+                loading: false,
             });
 
             await act(() =>
@@ -551,12 +557,6 @@ describe("SessionAuth", () => {
                 },
             };
 
-            // when
-            MockSession.validateClaims.mockImplementationOnce(({ userContext }) => {
-                saveInvalidClaimRedirectPathInContext(userContext, "/test-redirect");
-                return [{ id: "st-test-claim" }];
-            });
-
             await act(() =>
                 listenerFn({
                     action: "ACCESS_TOKEN_PAYLOAD_UPDATED",
@@ -596,12 +596,6 @@ describe("SessionAuth", () => {
 
             expect(await result.findByText(/^testClaimValue:/)).toHaveTextContent(`testClaimValue: undefined`);
 
-            // when
-            MockSession.validateClaims.mockImplementationOnce(({ userContext }) => {
-                saveInvalidClaimRedirectPathInContext(userContext, "/test-redirect");
-                return [{ id: "st-test-claim" }];
-            });
-
             await Promise.all([
                 act(() =>
                     listenerFn({
@@ -639,12 +633,6 @@ describe("SessionAuth", () => {
             expect(accessTokenPayloadElement).toHaveTextContent(`accessTokenPayload: ${JSON.stringify({})}`);
 
             expect(await result.findByText(/^testClaimValue:/)).toHaveTextContent(`testClaimValue: undefined`);
-
-            // when
-            MockSession.validateClaims.mockImplementationOnce(({ userContext }) => {
-                saveInvalidClaimRedirectPathInContext(userContext, "/test-redirect");
-                return [{ id: "st-test-claim" }];
-            });
 
             await Promise.all([
                 act(() =>